packages/media-core/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-core"
version = "0.1.0"
description = "Shared media processing primitives for Reframe"
authors = [{ name = "Reframe Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.7",
]

[project.optional-dependencies]
transcribe-openai = ["openai>=1.25.0"]
transcribe-faster-whisper = ["faster-whisper>=1.0.0"]
transcribe-whispercpp = ["whispercpp>=0.0.9"]
render = ["moviepy>=1.0.3", "pysubs2>=1.6.1"]
translate-local = ["argostranslate>=1.9.0"]
diarize-pyannote = ["pyannote.audio>=3.1.1"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
